requirements.txt
numpy>=1.21.2
razdel>=0.5.0
scipy>=1.7.3
sentencepiece>=0.1.95
six>=1.16.0
tensorflow-gpu>=2.5.0
tensorflow_addons>=0.13.0
tqdm>=4.62.3
tokenizers>=0.10.1
transformers>=4.5.1
